Refactored tests to match file split
